Make the surl.metrics helper command work
This is a bit of a hack, since `surl` parses the name of the snap
as if it were the URL. This commit makes two changes:

1. Change get_environment_from_url() to return a default
("staging", "snapcraft") tuple in cases where it fails to
recognize the URL that's passed in. In this case, the function
fails to recognize the name of a snap as a URL, and so it defaults
to returning ("staging", "snapcraft"). This is fine, as the values
get ovewritten later in __init__.py by the values passed in to the
command line.

2. Since get_config_from_cli() parses the name of the snap as the
URL, the parsed "url" is inserted into the front of the list of
arguments not parsed by get_config_from_cli() ("remainder") so
that the parser defined in surl_metrics.py can properly parse out
the "snap_name" argument.